Debugging Incident Insertion
Tried debugging where the issue is with inserting incidents into the tree. I have notes on line 268 with what the issues could be. I've been able to get it to work for the first incident but not for subsequent incidents. An error in reformat date pops up, which is why I think its an issue with either loop indexes of how we're reading the file.